chore(sign-in): update footer text in sign-in layout component
The footer text in the sign-in layout component has been updated from "NextUI" to "Mazmiiskndr" to reflect the correct branding.

chore(sign-in): adjust transition duration in sign-in page component

The transition duration in the sign-in page component has been adjusted from 0.5 seconds to 0.8 seconds to improve the visual experience.

chore(marketing): update footer text in marketing layout component

The footer text in the marketing layout component has been updated from "NextUI" to "Mazmiiskndr" to reflect the correct branding.

chore(marketing): add motion animation to marketing page component

The marketing page component now includes motion animation using the framer-motion library. The page has an initial opacity of 0, transitions to an opacity of 1 with a duration of 0.8 seconds when entering, and transitions back to an opacity of 0 with a duration of 0.8 seconds when exiting.